Refactor cfg.substitution package
